+++ wasm/no_std/src/lib.rs
@@ -1,6 +1,7 @@
 #![no_std]
 
-pub use key_share::*;
+pub use cggmp21_keygen;
+pub use key_share;
 
 #[panic_handler]
 fn panic(_panic: &core::panic::PanicInfo<'_>) -> ! {