+++ cggmp21-keygen/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0
+* Make library `#![no_std]`-compatible and WASM-friendly [#100]
+* Provide sync API to carry out DKG protocol [#100]
+* Update `round-based` dep to `v0.3`
+
+[#100]: https://github.com/dfns/cggmp21/pull/100
+
 ## v0.1.0
 
 Initial release